refactor(ui): réecrit les composants des pages header et footer en tsx
